More input file variety, inputs mod 1000
Allows the input file to be .asm (which is the standard for assembly
files, so it makes sense to support those), and for the extension to be
in any case (eg file.TXT).

Takes any inputs to the program (including those from the command line)
as modulo 1000, like the actual LMC does.