Added function for all test cases
I thought this had closed an issue, but rereading it there is one more
feature. I'll add that now.

Also added the -q (for quiet) flag to avoid filling the terminal when
using -a. Also added the -f flag, allowing the user to specify the
output file (relative to the input file) or to decide not to have one at
all.